pyproject.toml
# 1️⃣ pyproject.toml – dependencies, formatting & linting
[tool.poetry]
name = "loan_dash"
version = "0.1.0"
description = "Gold-loan KPI & visualisation dashboards"
authors = ["Ishan Kukade"]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.35"
pandas = "^2.2"
plotly = "^5.22"
sqlalchemy = "^2.0"
mysql-connector-python = "^8.4"
pydantic = "^2.7"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-mock = "^3.14"
black = "^24.4"
ruff = "^0.4"
isort = "^5.13"

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
